pyproject.toml
[project]
name = "dspy-rise-humbench"
version = "0.1.0"
description = "DSPy optimization pipeline for RISE Library Cards Benchmark"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy>=2.6",
    "python-dotenv",
    "rapidfuzz",
    "pydantic",
    "Pillow",
    "openai",
]
